+++ R/newMultispeciesParams.R
@@ -149,6 +149,7 @@ newMultispeciesParams <- function(
     collect_info <- function(cnd) {
         if (cnd$level <= info_level) {
             infos[[cnd$var]] <<- cnd$message
+            rlang::cnd_muffle(cnd)
         }
     }
     # Register this signal handler
@@ -186,6 +187,7 @@ newMultispeciesParams <- function(
         setParams(
                   # setInteraction
                   interaction = interaction,
+                  info_level = 0,
                   # setPredKernel()
                   pred_kernel = pred_kernel,
                   # setSearchVolume()
@@ -330,11 +332,24 @@ newMultispeciesParams <- function(
 #' @family functions for setting parameters
 # The reason we list `interaction` explicitly rather than including it in
 # the `...` is for backwards compatibility. It used to be the second argument.
-setParams <- function(object, interaction = NULL, ...) {
+setParams <- function(object, interaction = NULL, info_level = 3, ...) {
     UseMethod("setParams")
 }
 #' @export
-setParams.MizerParams <- function(object, interaction = NULL, ...) {
+setParams.MizerParams <- function(object, interaction = NULL,
+                                  info_level = 3, ...) {
+    # Define a signal handler that collects the information signals
+    # into the `infos` list.
+    infos <- list()
+    collect_info <- function(cnd) {
+        if (cnd$level <= info_level) {
+            infos[[cnd$var]] <<- cnd$message
+            rlang::cnd_muffle(cnd)
+        }
+    }
+    # Register this signal handler
+    withCallingHandlers(
+        info_about_default = collect_info, {
     params <- validParams(object)
 
     params <- setInteraction(params, interaction)
@@ -361,5 +376,9 @@ setParams.MizerParams <- function(object, interaction = NULL, ...) {
     }
 
     validObject(params)
+    })
+    if (length(infos) > 0) {
+        message(paste(infos, collapse = "\n"))
+    }
     params
 }